Move play detail components inside profile page
Nothing else is using it.
