+++ resources/assets/lib/profile-page/play-detail.tsx
